Keep track of time stamp of received data
This allows clients to detect, once collection is stopped, whether all data has been received yet or not. It also removes the assumption that data will always be appended. Now, stored data is always updated to match the sent data when the timestamp of the received data is newer than that of the stored data.